requirements.txt
pip
autopep8
Flask==3.0.3
PyMuPDF==1.24.2
python-docx==1.1.0
openai==1.13.3
pytesseract==0.3.10
gunicorn==22.0.0
Werkzeug==3.0.6